Regretion fixes Implemented a smarter clone if necessary algorithm

+++ openmole/core/org.openmole.core.batchservicecontrol/src/main/scala/org/openmole/core/batchservicecontrol/IQualityControl.scala
@@ -21,6 +21,7 @@ import org.openmole.core.batchservicecontrol.internal.Activator
 import org.openmole.core.model.execution.batch.BatchServiceDescription
 
 object IQualityControl {
+
   def withFailureControl[A](desc: BatchServiceDescription, op: => A): A = {
     val qualityControl = Activator.getRessourceControl.qualityControl(desc)
     try {
@@ -48,5 +49,5 @@ trait IQualityControl {
     def reinit
     def increaseQuality(value: Int)
     def decreaseQuality(value: Int)
-    def quality: Long
+    def quality: Int
 }
